refactor(Executor): Make Executor and interface; introduce BaseExecutor
